new firebase config setup
